+++ src/OsmSharp/IO/PBF/Encoder.cs
@@ -313,7 +313,7 @@ namespace OsmSharp.IO.PBF
                     : Encoder.EncodeTimestamp(previous.TimeStamp.Value, block.date_granularity);
                 groupDense.denseinfo.timestamp.Add(currentTimeStamp - previousTimeStamp);
                 groupDense.denseinfo.uid.Add((int)(current.UserId.Value - previous.UserId.Value));
-                groupDense.denseinfo.version.Add(current.Version.Value - previous.Version.Value);
+                groupDense.denseinfo.version.Add((int)(current.Version.Value - previous.Version.Value));
                 var previousUserNameId = previous.UserName == null
                     ? 0
                     : Encoder.EncodeString(block, reverseStringTable, previous.UserName);
